Convert setupModuleTest() to higher order factory function
