+++ lib/ember-mocha.js
@@ -1,9 +1,11 @@
 import describeModule    from 'ember-mocha/describe-module';
 import describeComponent from 'ember-mocha/describe-component';
 import describeModel     from 'ember-mocha/describe-model';
-import setupModuleTest   from 'ember-mocha/setup-module-test';
+import setupTestFactory  from 'ember-mocha/setup-test-factory';
 import { it }            from 'mocha';
-import { setResolver }   from 'ember-test-helpers';
+import { setResolver, TestModule } from 'ember-test-helpers';
+
+const setupModuleTest = setupTestFactory(TestModule);
 
 export {
   describeModule,
